+++ hypervisor/Makefile
@@ -213,6 +213,9 @@ HW_C_SRCS += arch/x86/notify.c
 HW_C_SRCS += arch/x86/vtd.c
 HW_C_SRCS += arch/x86/gdt.c
 HW_C_SRCS += arch/x86/irq.c
+HW_C_SRCS += arch/x86/irq_init.c
+HW_C_SRCS += arch/x86/exception.c
+HW_C_SRCS += arch/x86/nmi.c
 HW_C_SRCS += arch/x86/tsc.c
 HW_C_SRCS += arch/x86/tsc_deadline_timer.c
 HW_C_SRCS += common/timer.c
