+++ hypervisor/arch/x86/irq.c
@@ -216,51 +216,6 @@ void dispatch_interrupt(const struct intr_excp_ctx *ctx)
 	}
 }
 
-void dispatch_exception(struct intr_excp_ctx *ctx)
-{
-	uint16_t pcpu_id = get_pcpu_id();
-
-	/* Dump exception context */
-	dump_exception(ctx, pcpu_id);
-
-	/* Halt the CPU */
-	cpu_dead();
-}
-
-void handle_nmi(__unused struct intr_excp_ctx *ctx)
-{
-	uint32_t value32;
-
-	/*
-	 * There is a window where we may miss the current request in this
-	 * notification period when the work flow is as the following:
-	 *
-	 *       CPUx +                   + CPUr
-	 *            |                   |
-	 *            |                   +--+
-	 *            |                   |  | Handle pending req
-	 *            |                   <--+
-	 *            +--+                |
-	 *            |  | Set req flag   |
-	 *            <--+                |
-	 *            +------------------>---+
-	 *            |     Send NMI      |  | Handle NMI
-	 *            |                   <--+
-	 *            |                   |
-	 *            |                   |
-	 *            |                   +--> vCPU enter
-	 *            |                   |
-	 *            +                   +
-	 *
-	 * So, here we enable the NMI-window exiting to trigger the next vmexit
-	 * once there is no "virtual-NMI blocking" after vCPU enter into VMX non-root
-	 * mode. Then we can process the pending request on time.
-	 */
-	value32 = exec_vmread32(VMX_PROC_VM_EXEC_CONTROLS);
-	value32 |= VMX_PROCBASED_CTLS_NMI_WINEXIT;
-	exec_vmwrite32(VMX_PROC_VM_EXEC_CONTROLS, value32);
-}
-
 /* XXX lockless operation */
 bool irq_allocated_arch(struct irq_desc *desc)
 {
@@ -319,51 +274,5 @@ void init_irq_descs_arch(struct irq_desc descs[])
 		/* XXX irq0 -> vec0 ? */
 		irq_data[irq].vector = vr;
 		vector_to_irq[vr] = irq;
-	}
-}
-
-static void disable_pic_irqs(void)
-{
-	pio_write8(0xffU, 0xA1U);
-	pio_write8(0xffU, 0x21U);
-}
-
-static inline void fixup_idt(const struct host_idt_descriptor *idtd)
-{
-	uint32_t i;
-	struct idt_64_descriptor *idt_desc = idtd->idt->host_idt_descriptors;
-	uint32_t entry_hi_32, entry_lo_32;
-
-	for (i = 0U; i < HOST_IDT_ENTRIES; i++) {
-		entry_lo_32 = idt_desc[i].offset_63_32;
-		entry_hi_32 = idt_desc[i].rsvd;
-		idt_desc[i].rsvd = 0U;
-		idt_desc[i].offset_63_32 = entry_hi_32;
-		idt_desc[i].high32.bits.offset_31_16 = entry_lo_32 >> 16U;
-		idt_desc[i].low32.bits.offset_15_0 = entry_lo_32 & 0xffffUL;
-	}
-}
-
-static inline void set_idt(struct host_idt_descriptor *idtd)
-{
-	asm volatile ("   lidtq %[idtd]\n" :	/* no output parameters */
-		      :		/* input parameters */
-		      [idtd] "m"(*idtd));
-}
-
-void init_interrupt_arch(uint16_t pcpu_id)
-{
-	struct host_idt_descriptor *idtd = &HOST_IDTR;
-
-	if (pcpu_id == BSP_CPU_ID) {
-		fixup_idt(idtd);
-	}
-	set_idt(idtd);
-	init_lapic(pcpu_id);
-
-	if (pcpu_id == BSP_CPU_ID) {
-		/* we use ioapic only, disable legacy PIC */
-		disable_pic_irqs();
-		ioapic_setup_irqs();
 	}
 }
